+++ airbyte-integrations/connectors/source-okta/unit_tests/test_source.py
@@ -16,111 +16,106 @@ from source_okta.source import (
     UserRoleAssignments,
     Users,
 )
-from unit_tests import (
-    auth_token_config,
-    oauth_config,
-    token_config,
-    wrong_oauth_config_bad_auth_type,
-    wrong_oauth_config_bad_credentials_record,
-)
 
 
 class TestAuthentication:
-    def test_init_token_authentication_init(self):
-        config = token_config()
+    def test_init_token_authentication_init(self, token_config, auth_token_config):
+        config = token_config
         source_okta = SourceOkta()
         token_authenticator_instance = source_okta.initialize_authenticator(config=config)
         assert isinstance(token_authenticator_instance, TokenAuthenticator)
 
-        config = auth_token_config()
+        config = auth_token_config
         token_authenticator_instance = source_okta.initialize_authenticator(config=config)
         assert isinstance(token_authenticator_instance, TokenAuthenticator)
 
-    def test_init_oauth2_authentication_init(self):
-        config = oauth_config()
+    def test_init_oauth2_authentication_init(self, oauth_config):
+        config = oauth_config
         source_okta = SourceOkta()
         oauth_authentication_instance = source_okta.initialize_authenticator(config=config)
         assert isinstance(oauth_authentication_instance, OktaOauth2Authenticator)
 
-    def test_init_oauth2_authentication_wrong_credentials_record(self):
-        config = wrong_oauth_config_bad_credentials_record()
+    def test_init_oauth2_authentication_wrong_credentials_record(self, wrong_oauth_config_bad_credentials_record):
+        config = wrong_oauth_config_bad_credentials_record
         source_okta = SourceOkta()
         try:
             source_okta.initialize_authenticator(config=config)
         except Exception as e:
             assert e.args[0] == "Config validation error. `credentials` not specified."
 
-    def test_init_oauth2_authentication_wrong_oauth_config_bad_auth_type(self):
-        config = wrong_oauth_config_bad_auth_type()
+    def test_init_oauth2_authentication_wrong_oauth_config_bad_auth_type(self, wrong_oauth_config_bad_auth_type):
+        config = wrong_oauth_config_bad_auth_type
         source_okta = SourceOkta()
         try:
             source_okta.initialize_authenticator(config=config)
         except Exception as e:
             assert e.args[0] == "Config validation error. `auth_type` not specified."
 
-    def test_check_connection_ok(self, requests_mock):
-        config = oauth_config()
+    def test_check_connection_ok(self, requests_mock, oauth_config, api_url):
+        config = oauth_config
         source_okta = SourceOkta()
         oauth_authentication_instance = source_okta.initialize_authenticator(config=config)
         assert isinstance(oauth_authentication_instance, OktaOauth2Authenticator)
 
-        requests_mock.get("https://test_domain.okta.com/api/v1/users?limit=1", json={"connect": "ok"})
-        requests_mock.post("https://test_domain.okta.com/oauth2/v1/token", json={"access_token": "test_token", "expires_in": 948})
+        requests_mock.get(f"{api_url}/api/v1/users?limit=1", json={"connect": "ok"})
+        requests_mock.post(f"{api_url}/oauth2/v1/token", json={"access_token": "test_token", "expires_in": 948})
         assert source_okta.check_connection(logger=MagicMock(), config=config) == (True, None)
 
-    def test_check_connection_error_status_code(self, requests_mock):
-        config = oauth_config()
+    def test_check_connection_error_status_code(self, requests_mock, oauth_config, api_url):
+        config = oauth_config
         source_okta = SourceOkta()
         oauth_authentication_instance = source_okta.initialize_authenticator(config=config)
         assert isinstance(oauth_authentication_instance, OktaOauth2Authenticator)
 
-        requests_mock.get("https://test_domain.okta.com/api/v1/users?limit=1", status_code=400, json={})
-        requests_mock.post("https://test_domain.okta.com/oauth2/v1/token", json={"access_token": "test_token", "expires_in": 948})
+        requests_mock.get(f"{api_url}/api/v1/users?limit=1", status_code=400, json={})
+        requests_mock.post(f"{api_url}/oauth2/v1/token", json={"access_token": "test_token", "expires_in": 948})
 
         assert source_okta.check_connection(logger=MagicMock(), config=config) == (False, {})
 
-    def test_check_connection_error_with_exception(self, requests_mock):
-        config = oauth_config()
+    def test_check_connection_error_with_exception(
+        self, requests_mock, oauth_config, api_url, error_failed_to_authorize_with_provided_credentials
+    ):
+        config = oauth_config
         source_okta = SourceOkta()
         oauth_authentication_instance = source_okta.initialize_authenticator(config=config)
         assert isinstance(oauth_authentication_instance, OktaOauth2Authenticator)
 
-        requests_mock.get("https://test_domain.okta.com/api/v1/users?limit=1", status_code=400, json="ss")
-        requests_mock.post("https://test_domain.okta.com/oauth2/v1/token", json={"access_token": "test_token", "expires_in": 948})
+        requests_mock.get(f"{api_url}/api/v1/users?limit=1", status_code=400, json="ss")
+        requests_mock.post(f"{api_url}/oauth2/v1/token", json={"access_token": "test_token", "expires_in": 948})
 
         assert source_okta.check_connection(logger=MagicMock(), config="wrong_config") == (
             False,
-            "Failed to authenticate with the provided credentials",
+            error_failed_to_authorize_with_provided_credentials,
         )
 
-    def test_check_streams(self, requests_mock):
-        config = oauth_config()
+    def test_check_streams(self, requests_mock, oauth_config, api_url):
+        config = oauth_config
         source_okta = SourceOkta()
         oauth_authentication_instance = source_okta.initialize_authenticator(config=config)
         assert isinstance(oauth_authentication_instance, OktaOauth2Authenticator)
 
-        requests_mock.get("https://test_domain.okta.com/api/v1/users?limit=1", json={"connect": "ok"})
-        requests_mock.post("https://test_domain.okta.com/oauth2/v1/token", json={"access_token": "test_token", "expires_in": 948})
+        requests_mock.get(f"{api_url}/api/v1/users?limit=1", json={"connect": "ok"})
+        requests_mock.post(f"{api_url}/oauth2/v1/token", json={"access_token": "test_token", "expires_in": 948})
         streams = source_okta.streams(config=config)
         for i, _ in enumerate([Groups, Logs, Users, GroupMembers, CustomRoles, UserRoleAssignments, GroupRoleAssignments]):
             assert isinstance(streams[i], _)
 
-    def test_oauth2_refresh_token_ok(self, requests_mock):
-        config = oauth_config()
+    def test_oauth2_refresh_token_ok(self, requests_mock, oauth_config, api_url):
+        config = oauth_config
         source_okta = SourceOkta()
         oauth_authentication_instance = source_okta.initialize_authenticator(config=config)
         assert isinstance(oauth_authentication_instance, OktaOauth2Authenticator)
-        requests_mock.post("https://test_domain.okta.com/oauth2/v1/token", json={"access_token": "test_token", "expires_in": 948})
+        requests_mock.post(f"{api_url}/oauth2/v1/token", json={"access_token": "test_token", "expires_in": 948})
         result = oauth_authentication_instance.refresh_access_token()
         assert result == ("test_token", 948)
 
-    def test_oauth2_refresh_token_failed(self, requests_mock):
-        config = oauth_config()
+    def test_oauth2_refresh_token_failed(self, requests_mock, oauth_config, api_url, error_while_refreshing_access_token):
+        config = oauth_config
         source_okta = SourceOkta()
         oauth_authentication_instance = source_okta.initialize_authenticator(config=config)
         assert isinstance(oauth_authentication_instance, OktaOauth2Authenticator)
-        requests_mock.post("https://test_domain.okta.com/oauth2/v1/token", json={"token": "test_token", "expires_in": 948})
+        requests_mock.post(f"{api_url}/oauth2/v1/token", json={"token": "test_token", "expires_in": 948})
         try:
             oauth_authentication_instance.refresh_access_token()
         except Exception as e:
-            assert e.args[0] == "Error while refreshing access token: 'access_token'"
+            assert e.args[0] == error_while_refreshing_access_token
